Extract category data to separate file
