+++ src/components/Categories/data.ts
@@ -1,0 +1,14 @@
+type Category = {
+  id: string;
+  name: string;
+};
+
+export const categories: Category[] = [
+  { name: "Campus", id: "campus" },
+  { name: "News & Politics", id: "news-and-politics" },
+  { name: "Lifestyle", id: "lifestyle" },
+  { name: "Science & Tech", id: "science-and-tech" },
+  { name: "Art", id: "art" },
+  { name: "Features", id: "features" },
+  { name: "Sports", id: "sports" },
+];
