+++ src/components/Categories/index.tsx
@@ -3,20 +3,7 @@ import { styled } from "../../stitches.config";
 
 import { TextLink } from "../Link";
 
-type Category = {
-  id: string;
-  name: string;
-};
-
-const categories: Category[] = [
-  { name: "Campus", id: "campus" },
-  { name: "News & Politics", id: "news-and-politics" },
-  { name: "Lifestyle", id: "lifestyle" },
-  { name: "Science & Tech", id: "science-and-tech" },
-  { name: "Art", id: "art" },
-  { name: "Features", id: "features" },
-  { name: "Sports", id: "sports" },
-];
+import { categories } from "./data";
 
 const Bar = styled("div", {
   background: "$accentDark",
